// frontend/app/controller/model/workflow/stepPresetDTO.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BaseDto } from 'app/zynerator/dto/BaseDto.model';
import { UserDestinataireDTO } from './userDestinataireDTO';

export class StepPresetDTO extends BaseDto { 
    // id?: number;
    title?: string;
    level?: number;
    description?: string;
    workflowPresetId?: number;
    destinataires?: Array<UserDestinataireDTO>;
    actions?: ACTION[];
    duration?: number;
    addPV?: boolean;
    


    constructor() {

        super();
        this.title="";
        this.description ="";
        this.level= 1;
        this.workflowPresetId = 0;
        this.destinataires = new Array<UserDestinataireDTO>;
        this.duration = 0;
        this.addPV = false;

    }
    




}
export namespace StepPresetDTO {
    export type ActionsEnum = 'CREATE' | 'UPDATE' | 'DELETE';
    export const ActionsEnum = {
        CREATE: 'CREATE' as ActionsEnum,
        UPDATE: 'UPDATE' as ActionsEnum,
        DELETE: 'DELETE' as ActionsEnum
    };
}
export enum ACTION {
    SIGN = 'SIGN',
    APPROVE = 'APPROVE',
    REJECT = 'REJECT',
    PARAPHER = 'PARAPHER',
    PRESIGNER = 'PRESIGNER',
    ENVOI_COURRIER = 'ENVOI_COURRIER',
  }